+++ cpukit/libnetworking/rtems/rtems_syscall.c
@@ -689,13 +689,13 @@ rtems_bsdnet_close (rtems_libio_t *iop)
 }
 
 static ssize_t
-rtems_bsdnet_read (rtems_libio_t *iop, void *buffer, uint32_t   count)
+rtems_bsdnet_read (rtems_libio_t *iop, void *buffer, size_t count)
 {
 	return recv (iop->data0, buffer, count, 0);
 }
 
 static ssize_t
-rtems_bsdnet_write (rtems_libio_t *iop, const void *buffer, uint32_t   count)
+rtems_bsdnet_write (rtems_libio_t *iop, const void *buffer, size_t count)
 {
 	return send (iop->data0, buffer, count, 0);
 }
